isolate provider into seperate class to prevent SPI failure if JAI missing
