+++ modules/library/referencing/src/main/java/org/geotools/referencing/operation/transform/WarpTransform2D.java
@@ -2,7 +2,7 @@
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
  *
- *    (C) 2005-2008, Open Source Geospatial Foundation (OSGeo)
+ *    (C) 2005-2016, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public
@@ -19,6 +19,7 @@ package org.geotools.referencing.operation.transform;
 import java.awt.geom.Point2D;
 import java.awt.geom.Rectangle2D;
 import java.io.Serializable;
+import java.lang.reflect.Array;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
@@ -34,25 +35,17 @@ import javax.media.jai.WarpGeneralPolynomial;
 import javax.media.jai.WarpPolynomial;
 import javax.media.jai.WarpQuadratic;
 
-import org.geotools.metadata.iso.citation.Citations;
-import org.geotools.parameter.DefaultParameterDescriptor;
 import org.geotools.parameter.Parameter;
 import org.geotools.parameter.ParameterGroup;
-import org.geotools.referencing.NamedIdentifier;
-import org.geotools.referencing.operation.MathTransformProvider;
 import org.geotools.resources.XArray;
 import org.geotools.resources.i18n.Vocabulary;
 import org.geotools.resources.i18n.VocabularyKeys;
 import org.geotools.util.Utilities;
-import org.opengis.parameter.ParameterDescriptor;
 import org.opengis.parameter.ParameterDescriptorGroup;
-import org.opengis.parameter.ParameterNotFoundException;
 import org.opengis.parameter.ParameterValue;
 import org.opengis.parameter.ParameterValueGroup;
-import org.opengis.referencing.operation.MathTransform;
 import org.opengis.referencing.operation.MathTransform2D;
 import org.opengis.referencing.operation.NoninvertibleTransformException;
-import org.opengis.referencing.operation.Transformation;
 
 
 /**
@@ -396,7 +389,7 @@ public class WarpTransform2D extends AbstractMathTransform implements MathTransf
     @Override
     public ParameterDescriptorGroup getParameterDescriptors() {
         if (warp instanceof WarpPolynomial) {
-            return Provider.PARAMETERS;
+            return WarpTransform2DProvider.PARAMETERS;
         } else {
             return super.getParameterDescriptors();
         }
@@ -411,14 +404,14 @@ public class WarpTransform2D extends AbstractMathTransform implements MathTransf
             final WarpPolynomial poly = (WarpPolynomial) warp;
             final ParameterValue[] p = new ParameterValue[7];
             int c = 0;
-            p[c++] = new Parameter(Provider.DEGREE,   Integer.valueOf(poly.getDegree()));
-            p[c++] = new Parameter(Provider.X_COEFFS, poly.getXCoeffs());
-            p[c++] = new Parameter(Provider.Y_COEFFS, poly.getYCoeffs());
+            p[c++] = new Parameter(WarpTransform2DProvider.DEGREE,   Integer.valueOf(poly.getDegree()));
+            p[c++] = new Parameter(WarpTransform2DProvider.X_COEFFS, poly.getXCoeffs());
+            p[c++] = new Parameter(WarpTransform2DProvider.Y_COEFFS, poly.getYCoeffs());
             float s;
-            if ((s=poly.getPreScaleX ()) != 1) p[c++] = new Parameter(Provider. PRE_SCALE_X, s);
-            if ((s=poly.getPreScaleY ()) != 1) p[c++] = new Parameter(Provider. PRE_SCALE_Y, s);
-            if ((s=poly.getPostScaleX()) != 1) p[c++] = new Parameter(Provider.POST_SCALE_X, s);
-            if ((s=poly.getPostScaleY()) != 1) p[c++] = new Parameter(Provider.POST_SCALE_Y, s);
+            if ((s=poly.getPreScaleX ()) != 1) p[c++] = new Parameter(WarpTransform2DProvider. PRE_SCALE_X, s);
+            if ((s=poly.getPreScaleY ()) != 1) p[c++] = new Parameter(WarpTransform2DProvider. PRE_SCALE_Y, s);
+            if ((s=poly.getPostScaleX()) != 1) p[c++] = new Parameter(WarpTransform2DProvider.POST_SCALE_X, s);
+            if ((s=poly.getPostScaleY()) != 1) p[c++] = new Parameter(WarpTransform2DProvider.POST_SCALE_Y, s);
             return new ParameterGroup(getParameterDescriptors(), XArray.resize(p, c));
         } else {
             return super.getParameterValues();
@@ -601,116 +594,6 @@ public class WarpTransform2D extends AbstractMathTransform implements MathTransf
         @Override
         public PointDouble clone() {
             return this;
-        }
-    }
-
-    /**
-     * The provider for the {@link WarpTransform2D}. This provider constructs a JAI
-     * {@linkplain WarpPolynomial image warp} from a set of polynomial coefficients,
-     * and wrap it in a {@link WarpTransform2D} object.
-     *
-     * @version $Id$
-     * @author Martin Desruisseaux
-     */
-    public static class Provider extends MathTransformProvider {
-        /** Serial number for interoperability with different versions. */
-        private static final long serialVersionUID = -7949539694656719923L;
-
-        /** Descriptor for the "{@link WarpPolynomial#getDegree degree}" parameter value. */
-        public static final ParameterDescriptor<Integer> DEGREE = DefaultParameterDescriptor.create(
-                "degree", 2, 1, MAX_DEGREE);
-
-        /** Descriptor for the "{@link WarpPolynomial#getXCoeffs xCoeffs}" parameter value. */
-        public static final ParameterDescriptor X_COEFFS = new DefaultParameterDescriptor(
-                "xCoeffs", float[].class, null, null);
-
-        /** Descriptor for the "{@link WarpPolynomial#getYCoeffs yCoeffs}" parameter value. */
-        public static final ParameterDescriptor Y_COEFFS = new DefaultParameterDescriptor(
-                "yCoeffs", float[].class, null, null);
-
-        /** Descriptor for the "{@link WarpPolynomial#getPreScaleX preScaleX}" parameter value. */
-        public static final ParameterDescriptor PRE_SCALE_X;
-
-        /** Descriptor for the "{@link WarpPolynomial#getPreScaleY preScaleY}" parameter value. */
-        public static final ParameterDescriptor PRE_SCALE_Y;
-
-        /** Descriptor for the "{@link WarpPolynomial#getPostScaleX postScaleX}" parameter value. */
-        public static final ParameterDescriptor POST_SCALE_X;
-
-        /** Descriptor for the "{@link WarpPolynomial#getPostScaleY postScaleY}" parameter value. */
-        public static final ParameterDescriptor<Float> POST_SCALE_Y;
-        static {
-            final Float ONE = 1f;
-             PRE_SCALE_X = DefaultParameterDescriptor.create( "preScaleX",null, Float.class, ONE, false);
-             PRE_SCALE_Y = DefaultParameterDescriptor.create( "preScaleY", null, Float.class, ONE, false);
-            POST_SCALE_X = DefaultParameterDescriptor.create("postScaleX", null, Float.class, ONE, false);
-            POST_SCALE_Y = DefaultParameterDescriptor.create("postScaleY", null, Float.class, ONE, false);
-        }
-
-        /**
-         * The parameters group.
-         */
-        static final ParameterDescriptorGroup PARAMETERS = createDescriptorGroup(new NamedIdentifier[] {
-                new NamedIdentifier(Citations.GEOTOOLS, "WarpPolynomial")
-            }, new ParameterDescriptor[] {
-                DEGREE, X_COEFFS, Y_COEFFS, PRE_SCALE_X, PRE_SCALE_Y, POST_SCALE_X, POST_SCALE_Y
-            });
-
-        /**
-         * Create a provider for warp transforms.
-         */
-        public Provider() {
-            super(2, 2, PARAMETERS);
-        }
-
-        /**
-         * Returns the operation type.
-         */
-        @Override
-        public Class<Transformation> getOperationType() {
-            return Transformation.class;
-        }
-
-        /**
-         * Creates a warp transform from the specified group of parameter values.
-         *
-         * @param  values The group of parameter values.
-         * @return The created math transform.
-         * @throws ParameterNotFoundException if a required parameter was not found.
-         */
-        protected MathTransform createMathTransform(final ParameterValueGroup values)
-                throws ParameterNotFoundException
-        {
-            final int      degree   =        intValue(DEGREE,   values);
-            final float[] xCoeffs   = (float[]) value(X_COEFFS, values);
-            final float[] yCoeffs   = (float[]) value(Y_COEFFS, values);
-            final float   preScaleX = scale( PRE_SCALE_X, values);
-            final float   preScaleY = scale( PRE_SCALE_Y, values);
-            final float  postScaleX = scale(POST_SCALE_X, values);
-            final float  postScaleY = scale(POST_SCALE_Y, values);
-            final Warp warp;
-            switch (degree) {
-                case 1:  warp = new WarpAffine           (xCoeffs, yCoeffs, preScaleX, preScaleY, postScaleX, postScaleY); break;
-                case 2:  warp = new WarpQuadratic        (xCoeffs, yCoeffs, preScaleX, preScaleY, postScaleX, postScaleY); break;
-                case 3:  warp = new WarpCubic            (xCoeffs, yCoeffs, preScaleX, preScaleY, postScaleX, postScaleY); break;
-                default: warp = new WarpGeneralPolynomial(xCoeffs, yCoeffs, preScaleX, preScaleY, postScaleX, postScaleY); break;
-            }
-            return new WarpTransform2D(warp, (Warp)null);
-        }
-
-        /**
-         * Returns the parameter value for the specified operation parameter.
-         *
-         * @param  param The parameter to look for.
-         * @param  group The parameter value group to search into.
-         * @return The requested parameter value, or {@code 1} if none.
-         */
-        private static float scale(final ParameterDescriptor param,
-                                   final ParameterValueGroup group)
-                throws ParameterNotFoundException
-        {
-            final Object value = value(param, group);
-            return (value!=null) ? ((Number) value).floatValue() : 1;
         }
     }
 }
